backend/requirements.txt
setuptools>=65.5.0
wheel>=0.38.0
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
httpx>=0.25.0
python-multipart>=0.0.6
websockets>=11.0.3
openai>=1.6.1
langchain>=0.0.335
faiss-cpu>=1.7.4
numpy>=1.25.2
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
# For improved RAG capabilities
chromadb>=0.4.18
# For Whisper transcription and audio processing
openai-whisper>=20231117
pydub>=0.25.1
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
scipy>=1.11.0
webrtcvad>=2.0.10
nltk>=3.8.1

# Document processing packages
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
pytesseract>=0.3.10
pdf2image>=1.16.3
PyMuPDF>=1.23.14
Pillow>=10.0.0